fix: Type error with YAML dates in frontmatter
When valid YAML dates or times are added to the page frontmatter, the
YAML parser will already convert them to datetime.date or
datetime.datetime objects.

Example:

    ---
    date: 2023-04-12
    ---

This resulted in #strptime raising a TypeError because datetime.date
isn't a string. This commit changes the plugin code to always accept a
native datetime.date or datetime.datetime, even if meta_time_format is
not set, and handle them without string parsing.

Close #48